% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_matrix.R
\name{abundance_matrix}
\alias{abundance_matrix}
\title{Return an abundance matrix based on summed isotopes}
\usage{
abundance_matrix(SummedIsotopes, IonGroup)
}
\arguments{
\item{SummedIsotopes}{(summed isotopes) The dataframe of summed_isotopes from sum_isotopes. Required.}

\item{IonGroup}{(character) Must be an ion group to subset data down to (a-c or x-z). Required.}
}
\value{
(abundance matrix) A data.frame where each row is a fragment and
each column is a summed abundance.
}
\description{
Returns a data.frame of class "abundance_matrix_isoforma" where each
row is a fragment, and each column is a proteoform summed abundance.
}
\examples{
\dontrun{
# Load summed isotopes 
SumIso <- readRDS(system.file("extdata", "SummedIsotopes_1to1to1.RDS", package = "isoforma"))

# Generate the matrix
abundance_matrix(SummedIsotopes = SumIso, IonGroup = "c")
}
}
