% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_spectrum_ptms_plot.R
\name{annotated_spectrum_ptms_plot}
\alias{annotated_spectrum_ptms_plot}
\title{Overlays multiple PTM fragment identifications on one spectra}
\usage{
annotated_spectrum_ptms_plot(SummedSpectra, IsoformaFragments)
}
\arguments{
\item{SummedSpectra}{(peak data) A data.table with a "M/Z" and "Intensity" column, preferably
calculated with the function sum_spectra. Required.}

\item{IsoformaFragments}{(matched peaks isoforma) A matched peaks isoforma object from fragments_per_ptm. Required.}
}
\value{
(plotly graphic) An overlay of each identified peak over a spectrum
}
\description{
Overlays multiple PTM fragment identifications on one spectra
}
\examples{
\dontrun{
# Load the summed spectra
SumPeaks <- readRDS(system.file("extdata", "SummedPeaks_1to1to1.RDS", package = "isoforma"))

# Load fragment data
Fragments <- readRDS(system.file("extdata", "Fragments_1to1to1.RDS", package = "isoforma"))

annotated_spectrum_ptms_plot(SummedSpectra = SumPeaks, IsoformaFragments = Fragments)
}
}
