% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_modifications_object.R
\name{create_modifications_object}
\alias{create_modifications_object}
\title{A function to generate a list of properly formatted proforma strings}
\usage{
create_modifications_object(Sequence, Modification)
}
\arguments{
\item{Sequence}{A valid protein sequence. It may start with "M." which gets removed. Required.}

\item{Modification}{An IsoForma modifications annotation written as
"PTM,Residue(Positions)\link{Number of Modifications}". Required.}
}
\description{
A function to generate a list of properly formatted proforma strings
}
\examples{
\dontrun{

# Single modification of a single PTM examples 
create_modifications_object("TRICITIES", "Methyl,I(3)[1]")
create_modifications_object("TRICITIES", "Methyl,I(3,5,7)[1]")

# Multiple modifications of a single PTM example
create_modifications_object("TRICITIES", "Methyl,I(3,5,7)[2]") 

# Multiple modifications with a fixed position
create_modifications_object("TRICITIES", "Methyl,I(3^,5,7)[2]") 

# Multiple modifications with two fixed positions 
create_modifications_object("TRICITIES", "Methyl,I(3^,5,7^)[2]")

# Multiple modifications of a single PTM with any "X" residue
create_modifications_object("TRICITIES", "Methyl,X(1,2,3,4,5,6,7,8,9)[2]")

# Multiple modifications with multiple PTMs examples 
create_modifications_object("TRICITIES", "Methyl,T(1)[1];Acetyl,X(2,4,9)[1]")
create_modifications_object("TRICITIES", "Methyl,T(1,2,3,4,5,6,7,8,9)[1];Acetyl,X(2,4,9)[1]")
create_modifications_object("TRICITIES", "Methyl,T(1^,2,3,4,5,6,7^,8,9)[3];Acetyl,X(2,4,9)[1]") 

}
}
