% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{.determine_ions}
\alias{.determine_ions}
\title{Select ions based on dissociation method}
\usage{
.determine_ions(activation_method)
}
\arguments{
\item{ActivationMethod}{A dissociation method of "HCD", "CID", or "ETD". Otherwise,
all ions will be returned. Required.}
}
\description{
Select ions based on dissociation method
}
