% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_heatmap.R
\name{ptm_heatmap}
\alias{ptm_heatmap}
\title{Generate a heatmap of monoisotopic fragments for each PTM}
\usage{
ptm_heatmap(IsoformaFragments)
}
\arguments{
\item{IsoformaFragments}{A matched peaks isoforma object from fragments_per_ptm. Required.}
}
\description{
Generate a heatmap of monoisotopic fragments for each PTM
}
\examples{
\dontrun{



}
}
