% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_isotopes.R
\name{sum_isotopes}
\alias{sum_isotopes}
\title{Sum isotopes within a matched_peaks object}
\usage{
sum_isotopes(IsoformaFragments)
}
\arguments{
\item{IsoformaFragments}{A matched peaks isoforma object from fragments_per_ptm. Required.}
}
\description{
Sum isotopes within a matched_peaks object
}
\details{
Intensities are divided by their respective charge states and summed together.
}
\examples{
\dontrun{


}
}
