% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_spectra.R
\name{sum_spectra}
\alias{sum_spectra}
\title{Function to sum spectrum based on scan number selections, or a list of pspecterlib::peak_data objects}
\usage{
sum_spectra(
  ScanMetadata = NULL,
  ScanNumbers = NULL,
  PeakDataList = NULL,
  PPMRound = 5,
  MinimumAbundance = 0.01
)
}
\arguments{
\item{ScanMetadata}{(scan_metadata object) Object of the scan_metadata class from get_scan_metadata
in pspecterlib. Required if PeakDataList is NULL.}

\item{ScanNumbers}{(numeric) Scan numbers contained in the ScanMetadata
object. Can be calculated by pull_scan_numbers. Required if PeakDataList is NULL.}

\item{PeakDataList}{(peak_data object) A list of peak_data scan objects. Required if ScanMetadata and
ScanNumbers are NULL.}

\item{PPMRound}{(numeric) Number to round to in PPM. Default is 5.}
}
\description{
Function to sum spectrum based on scan number selections, or a list of pspecterlib::peak_data objects
}
\details{
Sum spectra based on peak selection from the pspecterlib ScanMetadata object,
or provide a list of pspecterlib::peak_data objects to sum.
}
\examples{
\dontrun{

#################################################
## EXAMPLE 1: PULL FROM A SCAN METADATA OBJECT ##
#################################################
            
            
#######################################
## EXAMPLE 2: SUM PEAK DATA TOGETHER ##               
#######################################

# Load 3 peaks to sum 
PeakDataList <- list(
  readRDS(system.file("extdata", "PeakData_1to1to1_1.RDS", package = "isoforma")),
  readRDS(system.file("extdata", "PeakData_1to1to1_2.RDS", package = "isoforma")),
  readRDS(system.file("extdata", "PeakData_1to1to1_3.RDS", package = "isoforma"))
)

sum_spectra(PeakDataList = PeakDataList)

}
}
