% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mgf_simple.R
\name{write_mgf_simple}
\alias{write_mgf_simple}
\title{A simple MGF writer}
\usage{
write_mgf_simple(
  MZ,
  Intensity,
  Outpath,
  Mass = "Summed",
  Title = "MS/MS gfma zoospore H2B Ac",
  Pepmass = "29095.7",
  Charge = "10+",
  RT = "10",
  Scans = "1"
)
}
\arguments{
\item{MZ}{(numeric) Vector of mz values. Must be same length as intensity vector.}

\item{Intensity}{(numeric) Vector of intensity values. Must be same length as mz vector.}

\item{Outpath}{(character) Path to write file.}

\item{Mass}{(character) The mass type of the peptide. Default provided.}

\item{Title}{(character) Name for scan. Default provided.}

\item{Pepmass}{(character) The mass of the peptide. Default provided.}

\item{Charge}{(character) The charge of the peptide. Default provided.}

\item{RT}{(character) The retention time of the peptide. Default provided.}

\item{Scans}{(character) The scan number of the peptide. Default provided.}
}
\value{
Writes an mgf file to the specified location in outpath
}
\description{
Some tools require MGF files. This function was written to accommodate
those external tools and provide a simple MGF write.
}
\examples{
\dontrun{
write_mgf_simple(MZ = c(1,2,3), Intensity = c(1,2,3), Outpath = "~/Downloads/Test.MGF")
}
    
}
