% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_matrix.R
\name{abundance_matrix}
\alias{abundance_matrix}
\title{Return an abundance matrix based on summed isotopes}
\usage{
abundance_matrix(IonGroup, SummedIsotopes, .Threshold = NULL)
}
\arguments{
\item{IonGroup}{Must be an ion group to subset data down to (a-c or x-z). Required.}

\item{SummedIsotopes}{The dataframe of summed_isotopes from sum_isotopes. Required.}
}
\description{
Returns a data.frame of class "abundance_matrix_isoforma" where each
row is a fragment, and each column is a proteoform summed abundance.
}
