% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated_spectrum_ptms_plot.R
\name{annotated_spectrum_ptms_plot}
\alias{annotated_spectrum_ptms_plot}
\title{Overlays multiple PTM fragment identifications on one spectra}
\usage{
annotated_spectrum_ptms_plot(SummedSpectra, IsoformaFragments)
}
\arguments{
\item{SummedSpectra}{A data.table with a "M/Z" and "Intensity" column, preferably
calculated with the function sum_spectra. Required.}

\item{IsoformaFragments}{A matched peaks isoforma object from fragments_per_ptm. Required.}
}
\description{
Overlays multiple PTM fragment identifications on one spectra
}
\examples{
\dontrun{


}
}
