% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_proportions.R
\name{calculate_proportions}
\alias{calculate_proportions}
\title{Generate the proportions of fragments from an abundance matrix}
\usage{
calculate_proportions(AbundanceMatrix, IncludePlot = TRUE, top = 8)
}
\arguments{
\item{AbundanceMatrix}{An abundance matrix from abundance_matrix(). Required.}

\item{IncludePlot}{TRUE/FALSE to indicate whether a plot should also be returned. Default is TRUE.}
}
\description{
Generate the proportions of fragments from an abundance matrix
}
