% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments_per_ptm.R
\name{fragments_per_ptm}
\alias{fragments_per_ptm}
\title{Calculate the fragments per PTM from the modifications_isoforma object}
\usage{
fragments_per_ptm(
  Sequences,
  SummedSpectra,
  PrecursorCharge,
  ActivationMethod,
  CorrelationScore = 0.7,
  PPMThreshold = 10,
  IsotopeAlgorithm = "Rdisop",
  Messages = FALSE
)
}
\arguments{
\item{Sequences}{A vector of valid proforma sequences. Required.}

\item{SummedSpectra}{A data.table with a "M/Z" and "Intensity" column, preferably
calculated with the function sum_spectra. Required.}

\item{PrecursorCharge}{A single numeric to determine the Precursor Charge state
(maximum charge) to consider fragments in the calculation. Required.}

\item{ActivationMethod}{A string to determine ions. "HCD", "CID", and "ETD"
select b and y; a, b, and y; and c and z, respectively. Anything else
returns all 6 ions. Required.}

\item{CorrelationScore}{A minimum Correlation Score to filter peaks with at
least 2 isotopes. Default is 0.7. Optional.}

\item{PPMThreshold}{A minimum PPM Threshold for matching calculated and exprimental
peaks. Default is 10. Optional.}

\item{IsotopeAlgorithm}{"isopat" uses the isopat package to calculate isotopes,
while "Rdisop" uses the Rdisop package. Though more accurate, Rdisop has been
known to crash on Windows computers when called iteratively more than 1000 times.
Default is Rdisop, though isopat is an alternative.}

\item{Messages}{A TRUE/FALSE to indicate whether status messages should be printed.
Default is FALSE.}
}
\description{
Calculate the fragments per PTM from the modifications_isoforma object
}
