% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoforma_pipeline.R
\name{isoforma_pipeline}
\alias{isoforma_pipeline}
\title{Run the main isoforma function}
\usage{
isoforma_pipeline(parameters_path, mzML_path, output_path)
}
\arguments{
\item{parameters_path}{The path to the parameters file. Required.}

\item{mzML_path}{The path to the mzML file. Required.}

\item{output_path}{The path to output the isoforma results folder. Default is same as mzML_path.}
}
\description{
Run the main isoforma function
}
