% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoforma_pipeline.R
\name{isoforma_pipeline}
\alias{isoforma_pipeline}
\title{Calculates fragments, sums isotopes and charge states, generates the abundance matrix,
and calculates the proportions of each proteoform}
\usage{
isoforma_pipeline(
  Sequences,
  SummedSpectra,
  PrecursorCharge,
  ActivationMethod,
  IonGroup,
  CorrelationScore = 0.7,
  PPMThreshold = 10,
  IsotopeAlgorithm = "Rdisop",
  Messages = FALSE,
  Top = 8
)
}
\arguments{
\item{Sequences}{(character) A vector of valid ProForma sequences. Required.}

\item{SummedSpectra}{(peak_data) A data.table with a "M/Z" and "Intensity" column,
preferably calculated with the function sum_spectra. Required.}

\item{PrecursorCharge}{(numeric) A single numeric to determine the Precursor
Charge state (maximum charge) to consider fragments in the calculation. Required.}

\item{ActivationMethod}{(character) A string to determine ions. "HCD", "CID", and
"ETD" select b and y; a, b, and y; and c and z, respectively. Anything else
returns all 6 ions. Required.}

\item{IonGroup}{(character) Must be an ion group to subset data down to (a-c or x-z). Required.}

\item{CorrelationScore}{(numeric) A minimum Correlation Score to filter peaks with
at least 2 isotopes. Default is 0.7. Optional.}

\item{PPMThreshold}{(numeric) A minimum PPM Threshold for matching calculated
and exprimental peaks. Default is 10. Optional.}

\item{IsotopeAlgorithm}{(character) "isopat" uses the isopat package to calculate
isotopes, while "Rdisop" uses the Rdisop package. Though more accurate,
Rdisop has been known to crash on Windows computers when called iteratively
more than 1000 times. Default is Rdisop, though isopat is an alternative.}

\item{Top}{(integer) The top N proportions to return, ranked from higest to lowest. Default is 8.}

\item{Message}{(logic) A TRUE/FALSE to indicate whether status messages should be printed. Default is FALSE.}
}
\value{
(list) A list with 5 objects: matched peaks isoforma, summed isotopes isoforma,
abundance matrix isoforma, the proportions dataframe, and a ggplot of the proportions
}
\description{
A wrapper function to run fragments_per_ptm, sum_isotopes, abundance_matrix,
and calculate_proportions.
}
\examples{
\dontrun{

####################################
## EXAMPLE 1: SKIP PEAK SELECTION ##
####################################

# Load 3 peaks to sum 
PeakDataList <- list(
 readRDS(system.file("extdata", "PeakData_1to1to1_1.RDS", package = "isoforma")),
 readRDS(system.file("extdata", "PeakData_1to1to1_2.RDS", package = "isoforma")),
 readRDS(system.file("extdata", "PeakData_1to1to1_3.RDS", package = "isoforma"))
)

# Sum peaks 
PeakSum <- sum_ms2_spectra(PeakDataList = PeakDataList)

# Get proteoform strings
MultipleMods <- pspecterlib::multiple_modifications(
    Sequence = "LQIFVKTLTGKTITLEVEPSDTIENVKAKIQDKEGIPPDQQRLIFAGKQLEDGRTLSDYNIQKESTLHLVLRLRGG",
    Modification = "6.018427,V(17,26,70)[1]",
    ReturnUnmodified = TRUE
)

# Run main pipeline function
IsoForma <- isoforma_pipeline(
    Sequences = MultipleMods,
    SummedSpectra = PeakSum,
    PrecursorCharge = 11,
    ActivationMethod = "ETD",
    IonGroup = "c",
    CorrelationScore = 0,
    IsotopeAlgorithm = "Rdisop", # If this step crashes, switch to "isopat" 
    Messages = TRUE
)

#########################################
## EXAMPLE 2: AUTOMATED PEAK SELECTION ##
#########################################

}


}
