% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_isotopes.R
\name{sum_isotopes}
\alias{sum_isotopes}
\title{Sum isotopes within a matched_peaks object}
\usage{
sum_isotopes(IsoformaFragments, Charges = 1:100)
}
\arguments{
\item{IsoformaFragments}{A matched peaks isoforma object from fragments_per_ptm. Required.}

\item{Charges}{The charge values to sum. Default is 1-100.}
}
\description{
Sum isotopes within a matched_peaks object
}
\examples{
\dontrun{


}
}
