% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_spectra.R
\name{sum_spectra}
\alias{sum_spectra}
\title{Function to sum spectrum based on scan number}
\usage{
sum_spectra(ScanMetadata, ScanNumbers, MZRound = 3, Percentage = 25)
}
\arguments{
\item{ScanMetadata}{Object of the scan_metadata class from get_scan_metadata
in pspecterlib. Required.}

\item{ScanNumbers}{A vector of scan numbers contained in the ScanMetadata
object. Can be calculated by pull_scan_numbers. Required.}

\item{MZRound}{Decimal place the MZ values should be rounded to for binning.
Default is 3.}

\item{Percentage}{The percentage of peaks to take. Default is 25\%.}
}
\description{
Function to sum spectrum based on scan number
}
\examples{
\dontrun{

# Generate the ScanMetadata object 
ScanMetadata <- pspecterlib::get_scan_metadata(
    MSPath = "/Users/degn400/Desktop/IsoForma_Test/Sorghum-Histone0622162L11.mzML"
)

# Run function
sum_spectra(ScanMetdata = ScanMetadata,
            ScanNumbers = c(1236, 1237, 1240, 1316, 1497, 1520, 1534))
}
}
