% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mgf_simple.R
\name{write_mgf_simple}
\alias{write_mgf_simple}
\title{A simple MGF writer}
\usage{
write_mgf_simple(
  mz,
  intensity,
  outpath,
  mass = "Summed",
  title = "MS/MS gfma zoospore H2B Ac",
  pepmass = "29095.7",
  charge = "10+",
  rt = "10",
  scans = "1"
)
}
\arguments{
\item{mz}{Vector of mz values. Must be same length as intensity vector.}

\item{intensity}{Vector of intensity values. Must be same length as mz vector.}

\item{outpath}{Path to write file.}

\item{mass}{The mass type of the peptide. Default provided.}

\item{title}{Name for scan. Default provided.}

\item{pepmass}{The mass of the peptide. Default provided.}

\item{charge}{The charge of the peptide. Default provided.}

\item{rt}{The retention time of the peptide. Default provided.}

\item{scans}{The scan number of the peptide. Default provided.}
}
\description{
A simple MGF writer
}
